apps/investigation-graph/requirements.txt
databricks-sql-connector>=3.0.0
databricks-sdk>=0.20.0
